+++ karma.start.js
@@ -1,5 +1,9 @@
 // Setup global test variables
-var dsLocalStorageAdapter, User, datastore;
+var dsLocalStorageAdapter, Profile, User, Post, Comment, datastore;
+
+assert.equalObjects = function (a, b, m) {
+  assert.deepEqual(JSON.parse(JSON.stringify(a)), JSON.parse(JSON.stringify(b)), m || 'Objects should be equal!');
+};
 
 // Helper globals
 var fail = function (msg) {
@@ -43,6 +47,57 @@ beforeEach(function () {
 
   datastore = new JSData.DS();
 
-  User = datastore.defineResource('user');
+  Profile = datastore.defineResource({
+    name: 'profile'
+  });
+  User = datastore.defineResource({
+    name: 'user',
+    relations: {
+      hasMany: {
+        post: {
+          localField: 'posts',
+          foreignKey: 'post'
+        }
+      },
+      hasOne: {
+        profile: {
+          localField: 'profile',
+          localKey: 'profileId'
+        }
+      }
+    }
+  });
+  Post = datastore.defineResource({
+    name: 'post',
+    relations: {
+      belongsTo: {
+        user: {
+          localField: 'user',
+          localKey: 'userId'
+        }
+      },
+      hasMany: {
+        comment: {
+          localField: 'comments',
+          foreignKey: 'postId'
+        }
+      }
+    }
+  });
+  Comment = datastore.defineResource({
+    name: 'comment',
+    relations: {
+      belongsTo: {
+        post: {
+          localField: 'post',
+          localKey: 'postId'
+        },
+        user: {
+          localField: 'user',
+          localKey: 'userId'
+        }
+      }
+    }
+  });
   dsLocalStorageAdapter = new DSLocalStorageAdapter();
 });